requirements.txt
pytest
selenium
pytest-html
pytest-xdist
pytest-dependency
pytest-ordering
requests
openpyxl
allure-pytest